+++ build.gradle
@@ -4,26 +4,12 @@ plugins {
 }
 
 // Group and Version for Gradle Plugin
-group = 'de.pfannekuchen'
+group = 'com.minecrafttas'
 version = '1.0.2'
-
-java {
-    withJavadocJar()
-    withSourcesJar()
-}
-
-gradlePlugin {
-    plugins {
-        mcpgradleog {
-            id = 'mcpgradleog'
-            implementationClass = 'de.pfannekuchen.mcpgradle.MCPGradle'
-        }
-    }
-}
 
 repositories {
 	mavenCentral()
-	maven { url "https://maven.mgnet.work/" }
+	maven { url "https://maven.mgnet.work/main" }
 }
 
 dependencies {
@@ -32,4 +18,13 @@ dependencies {
 	implementation group: 'de.pfannekuchen.mcpgradle', name: 'mcinjector-og', version: '2.0' // For MCP
 	implementation group: 'de.pfannekuchen.mcpgradle', name: 'retroguard-og', version: '3.0.0' // For MCP
 	implementation group: 'de.pfannekuchen.mcpgradle', name: 'jadretro-og', version: '1.6' // For MCP
+}
+
+gradlePlugin {
+    plugins {
+        mcpgradleog {
+            id = 'mcpgradleog'
+            implementationClass = 'com.minecrafttas.mcp4gradle.MCP4Gradle'
+        }
+    }
 }
